requirements.txt
langchain>=0.1.17
langchain-core
langchain-community
langchain_google_genai
google-generativeai
faiss-cpu
fastapi
uvicorn[standard]
gradio
beautifulsoup4
python-dotenv
pydantic
requests
lxml